README: initial CLI design
In addition, set up some basic settings for working in VS Code (to be
expanded as Go code is added).

Also add a wishlist of features that will be delayed until after the
prototype is complete. While building the prototype, please expand upon
this list whenever inspiration strikes instead of bogging down the issue
tracker.