update: add incremental bundle on top of previous bundles
Be sure to parse the existing bundle headers in order to get the
previous refs, then use those as prerequisites for the new incremental
bundle. It's not enough to only use the refs from the previous layer,
since perhaps not all refs were updated or maybe a ref was force-pushed.