update: collapse list if it grows longer than 5 items
This constant of "5" will one day be replaced with a configurable value
whose default is probably "30". It has other issues when we add the
concept of "daily" and "hourly" bundles.

For now, 5 is sufficient. Convert the oldest bundles together into a
single bundle, now called "base-<token>.bundle". We need to be careful
when defining the refs to use for each of these tips, so create a new
'refs/base/' ref namespace. Some TODOs are marked to use 'git merge-base
--independent' to reduce the number of tips used by this base bundle.

We need to be careful, but it is possible that we could also "GC" the
base bundle if the tip commits are not reachable from any tips in the
later bundles. In such a case, we might need to rebuild all of those
bundles to ensure that the REF_DELTA bases exist as needed (perhaps this
is part of the prereq commits, so it might not be a real worry).