ci: remove explicit 'GOPATH' and add note about '-C' option
The 'git-bundle-server' repository is checked out to 'src/git-bundle-server'
in anticipation of future dependencies (e.g., Git) being checked out to
parallel 'src' directories for use in integration tests.

The initial implementation of this workflow included explicit specification
of the 'GOPATH' for both 'go build' and 'go vet'; this is not necessary, so
remove it. Additionally, add a note about how the upcoming '-C' option
should remove the need for an explicit '-C' in each build step.

Signed-off-by: Victoria Dye <vdye@github.com>