daemon: create daemon management interface
Create 'DaemonProvider' interface for creating and managing daemon
processes, to eventually be used by the 'web-server' subcommand to manage a
web server daemon.

Because daemon process management is dependent on the operating system/tools
available, initially create two types to implement the interface: one for
'launchd' (MacOS) and one for 'systemd' (Linux). The functions, for now, all
return "not implemented" errors but will be implemented in later patches.

Finally, add a 'NewDaemonProvider' to create the appropriate daemon provider
implementation based on the operating system.

Signed-off-by: Victoria Dye <vdye@github.com>