launchd: implement 'Create()' function
Implement and test the 'Create()' function for the 'launchd' implementation
of 'DaemonProvider'. To create the service, this function first generates
the 'plist' [1] defining service metadata (the name, program to execute,
etc), then writes the file and loads - but does not start - the service.

The exact behavior is dependent on the 'force' flag. If it is 'false',
'Create()' will not overwrite the existing service; if 'true', it unloads
the old service (if applicable), writes the new file, and reloads.

Finally, add unit tests covering various scenarios that may occur when
creating the 'launchd' service. Add mocks and other common structures &
functions to a dedicated 'shared_test.go' file, as they will be useful when
testing other daemon provider implementations in later patches.

[1] https://www.unix.com/man-page/osx/5/launchd.plist/

Signed-off-by: Victoria Dye <vdye@github.com>