launchd: implement 'Start()' function
Implement and test the 'Start()' function for the 'launchd' daemon process
manager. The 'Start()' function assumes that the specified service has
already been boostrapped (with 'Create()'), and simply runs 'launchctl
kickstart' to start the service. If the service is already running,
'Start()' is effectively a no-op (as 'launchctl kickstart' does not restart
the already-running process).

Signed-off-by: Victoria Dye <vdye@github.com>