subcommand: rename 'subcommand()' to 'Name()'
Change the 'Subcommand' interface's 'subcommand()' to 'Name()' to clarify
what that the function is intended to return the name of the subcommand (as
it would be used on the command line). Additionally, by capitalizing the
function name, make it public to other packages; this will be necessary in
later patches when the interface is moved to a new package.

Signed-off-by: Victoria Dye <vdye@github.com>