argparse: create basic argument parser
Add a basic argument parser for arbitrary series of string arguments. For
now, it is little more than a wrapper of 'flag.FlagSet' (a builtin structure
that handles parsing of flags preceded by a '-'). In future patches, the arg
parser will be extended to handle subcommands and positional arguments to
simplify 'git-bundle-server' command setup.

Signed-off-by: Victoria Dye <vdye@github.com>