argparse: add parsing for subcommands
Add handling for subcommands in the custom arg parser. To specify a
subcommand, a user can call 'parser.Subcommand' and provide an
'argparse.Subcommand' instance, whose 'Name()' is used to identify its use
on the command line. When the specified subcommand is identifies, its
'Run()' function can be called in 'InvokeSubcommand()'. This function, when
called, is provided the remaining arguments *after* the subcommand, to allow
the subcommand function to perform its own argument parsing.

To use the most accurate terminology in usage printouts, include a function
('SetIsTopLevel()') to toggle whether the subcommands listed in the printout
are called "Commands" or "Subcommands".

Signed-off-by: Victoria Dye <vdye@github.com>