daemon: support programs with arguments
Add support for specifying arguments in both launchd and systemd
applications. In launchd, these args are specified in a "ProgramArguments"
array; in systemd, they are simply appended to the program name (as someone
would run a program on the command line).

Note that "ProgramArguments" in the 'launchd' plist now always has at least
one element: the path to the executable program. This aligns 'launchd' to
standard argument conventions [1]; although it does not yet affect
'git-bundle-web-server' (the only usage of the 'launchd' code at this
point), it will in future patches when arguments & arg parsing are added.

Finally, add & update unit tests covering the use of arguments in both
launchd & systemd.

[1] https://man7.org/linux/man-pages/man3/execvp.3.html

Signed-off-by: Victoria Dye <vdye@github.com>