launchd: replace plist template with XML encoder
Replace the template-based creation of a 'plist' file (used to configure a
launchd daemon) with XML marshalling. There are a couple of benefits to
configuring the 'plist' with Go's XML infrastructure, mainly built-in
character escaping and avoiding the fragility of defining the structured XML
format as a plaintext string.

As part of this change, add more tests of 'plist' content, and make the
tests more thoroughly verify file contents ensuring:

- that the generated plist has all the expected options
- that the plist doesn't include more options than it should
- that characters are properly escaped

Signed-off-by: Victoria Dye <vdye@github.com>