git-bundle-web-server: handle starting & repeated '/' in route
Rather than step through the provided route character-by-character, split
the route into non-empty tokens divided by '/'. This avoids the need for
special handling of leading or trailing slashes, as well as treats multiple
sequential slashes as a single divider.

It's possible that the "multiple slashes" case is already handled by Go's
internal route handling, but this implementation gives us a bit of extra
safety & is generally a more concise implementation.

Signed-off-by: Victoria Dye <vdye@github.com>