git-bundle-web-server: allow serving over HTTPS
Add '--cert' and '--key' options to 'git-bundle-web-server' to specify paths
to a X.509 SSL certificate and private key (respectively) to be used to
configure HTTPS. If both options are specified, the web server is hosted
with 'ListenAndServeTLS()' using the given credentials; if both are not
specified, the server is hosted (as it is now) with 'ListenAndServe()'. If
only one of the two options is specified, an error (with usage) is thrown
and the server is not hosted.

Signed-off-by: Victoria Dye <vdye@github.com>