web-server: capture and use web server options
Allow 'git-bundle-server web-server' to specify command line arguments for
the 'git-bundle-web-server' daemon. Like 'git-bundle-web-server' itself, use
'utils.WebServerFlags()' to configure the options (with a usage string that
indicates that it is a web server flag) & validation. The results of the
options are added to 'DaemonConfig.Arguments', with '--cert' and '--key'
explicitly converted to absolute paths to ensure the daemon finds the files
correctly.

Signed-off-by: Victoria Dye <vdye@github.com>