repo: make GetRepositories() mockable
In an effort to make the codebase more unit-testable, we need to be able
to mock certain lower layers of the system. We already have a FileSystem
interface for interacting with the filesystem, but it is not yet robust
enough for us to parse an existing file, as needed for
GetRepositories().

Add a new ReadFileLines() method that takes a filename and returns a
string slice (or error) and adapt GetRepositories() to use it. Note that
this abstracts both os.OpenFile() and bufio.NewReader(). This should
simplify the interface for most content interaction, but also will
simplify the mocks we will eventually create for this method.

Adjust all callers of GetRepositories() to provide a FileSystem,
specifically the "real" one for now, until we can replace enough items
with mockable interfaces that those methods can use an injected
dependency at the command level.

Helped-by: Victoria Dye <vdye@github.com>
Signed-off-by: Derrick Stolee <derrickstolee@github.com>