paths: use user.User value to get home dir
Continuing our efforts to make our code unit testable, add a user.User
struct to the parameters for the methods in paths.go. This allows us to
remove the rest of the dependencies on the real system in the
GetRepositories() method. It requires some more work elsewhere in the
codebase for all the places where the path methods are used.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>