repo: test GetRepositories()
Create a unit test setup for the GetRepositories() method, to ensure
that it works correctly with mocks.

For this method, we only need to mock FileSystem.ReadFileLines() once,
so we can have our tests store exactly one list of lines.

Create a fake user.User struct to provide the HomeDir value.

The only test implemented right now is the empty file.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>