repo: add more GetRepositories() tests, fix bug
Add a few more cases to TestRepos_GetRepositories() including an error
from the filesystem, a single repo case, and multiple repositories.

Be sure to match the expected output of ReadFileLines() which is to trim
the newlines.

In the multiple repositories case, I added an empty line to check the
case of skipping newlines. In an earlier version of this commit, I had
thought there was a bug, but it was due to me including newlines in the
mocked output.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>