release.yml: create release pipeline
Create a basic release pipeline that triggers when a tag of the correct
format ("v<number>.<number>.<number>...") is pushed. The pipeline generates
three artifacts:

1. An x86/64 .deb binary package
2. An x86/64 MacOS .pkg
3. An ARM64 MacOS .pkg

At the moment, these packages are simply uploaded as artifacts in the
workflow. In the future, the goal is to automatically create and attach
these artifacts to a draft release, to allow maintainers to add release note
details then publish with the necessary attachments.

Note that the tag filtering in GitHub Actions isn't fancy enough to allow
for full verification of the version, so a tag like "v1a.2b.3c-xyz" would
trigger the pipeline. Better protections will be put in place later, but for
now we can rely on maintainers to push only valid tags for the pipeline.

Signed-off-by: Victoria Dye <vdye@github.com>