web-server: add '--remove' option to 'stop'
Add a '--remove' option to 'git-bundle-server web-server stop' to fully
remove the service daemon and its resources on disk after stopping the
daemon process. For launchd, this is done by first running 'launchctl
bootout' on the appropriate service, then removing the plist file; for
systemd, the service unit file is removed then 'systemctl daemon-reload' is
run. Neither implementation fails if the web server was never loaded.

Although it is unlikely that users will need this option in day-to-day use,
this cleanup option will be needed to fully remove bundle server resources
in the uninstall scripts implemented in later patches.

Signed-off-by: Victoria Dye <vdye@github.com>