Makefile: create basic 'build' and 'clean' targets
Popular convention in Go projects seems to be using Makefiles for automating
build/release tasks [1]. In anticipation of adding complex packaging scripts
in future patches, create an initial Makefile with basic 'build' and 'clean'
targets.

Note that the 'build' target utilizes the 'GOOS' and 'GOARCH' of the
environment to allow generating executables for any supported Go platform,
defaulting to the current platform the build is running on. This behavior
will be critical in later patches, when we want to create MacOS ARM64
packages on MacOS Intel build machines.

[1] https://github.com/golang-standards/project-layout/tree/master#scripts

Signed-off-by: Victoria Dye <vdye@github.com>