Makefile: add Debian packaging targets
Add 'package' target to Makefile that, based on the 'GOOS', builds the
platform-appropriate package. For now, support only binary .deb packages
(corresponding to a 'GOOS' of 'linux'); if the 'GOOS' is unsupported, 'make'
will exit with an error.

The .deb packaging process is broken into three steps, each corresponding to
a target in the Makefile:

1. Build the executables.
2. Copy the executables & any symlinks into the appropriate file structure
   for running 'dpkg-deb'.
3. Generate the control file, copy the maintainer script(s) [1], and build
   the package with 'dpkg-deb'.

Additionally, the 'check-arch' and 'check-version' targets verify that the
'GOARCH' is supported for packaging on the platform and that a version
string is set, respectively.

In accordance with the 'project-layout' guidelines for the Makefile [2],
most of the packaging logic outlined above lives in dedicated scripts. Best
practice for the location of these scripts is somewhat vague, since it
appears that both 'scripts/' [3] and 'build/package/' [4] would be valid
choices. Go with the latter, putting the scripts in 'build/package' (if they
will apply to more than one packaging mechanism) or a 'deb' subdirectory for
Debian-specific configuration.

[1] https://www.debian.org/doc/debian-policy/ap-flowcharts.html
[2] https://github.com/golang-standards/project-layout/tree/master#scripts
[3] https://github.com/golang-standards/project-layout/blob/master/scripts/README.md
[4] https://github.com/golang-standards/project-layout/blob/master/build/README.md

Signed-off-by: Victoria Dye <vdye@github.com>