Makefile: add code signing step for MacOS
Add a new script 'codesign.sh' to sign the contents of a given MacOS
directory with the appropriate Apple Developer ID certificate (assumed to
already be set up in a keychain). Add a 'codesign' target to the 'make
package' workflow that only runs if the 'APPLE_APP_IDENTITY' build variable
is set.

Additionally, add an '--identity' argument to 'pack.sh' to sign the .pkg
generated by 'productbuild' - if 'APPLE_INST_IDENTITY' is not set, the .pkg
is not signed.

Note that the signing identities for 'codesign.sh' and 'pack.sh' are
different variables: 'APPLE_APP_IDENTITY' is the application
signing identity, and 'APPLE_INST_IDENTITY' is the installer signing
identity. More information on this signing process can be found at [1] and
[2].

Finally, change the temporary package name suffix from '.tmp' to
'.component' to more clearly indicate that it is a component package.

[1] https://developer.apple.com/forums/thread/701514
[2] https://developer.apple.com/forums/thread/701581

Signed-off-by: Victoria Dye <vdye@github.com>