launchd: enable running after user logs out
Change the launch agent domain from 'gui/<uid>' (tied to login session) to
'user/<uid>' (tied to user, but may run when logged out). To allow the agent
to be loaded with this new domain, explicitly set the
'LimitLoadToSessionType' value in the generated plist to 'Background'
(default value seems to be 'Aqua', corresponding to the logged in session
i.e. 'gui/<uid>' domain).

Signed-off-by: Victoria Dye <vdye@github.com>