systemd: do not fail 'stop' if service unit not installed
If a service unit configuration file doesn't exist on disk, 'systemctl stop'
fails with error code 5 [1]. To keep 'git-bundle-server web-server stop'
from failing when 'web-server start' was not called before it, do not cause
the program to fail when the error code indicates the service isn't
installed. Add a test covering this case.

[1] https://freedesktop.org/software/systemd/man/systemd.exec.html#Process%20Exit%20Codes

Signed-off-by: Victoria Dye <vdye@github.com>