launchd: ignore error on already-stopped service
If a 'launchd' service is bootstrapped but not currently running, calling
'launchctl kill SIGINT' on the service will result in error code 3 ("No such
process"). Stopping an already-stopped service isn't an error case, so
update to ignore the error code.

As part of adding a new test to cover this error code, reorganize
'TestLaunchd_Stop' into table-driven tests to reduce code duplication.

Signed-off-by: Victoria Dye <vdye@github.com>