docs/man: add asciidoc manpage content
Add Asciidoc manual pages for 'git-bundle-server' (describing bundle servers
in general as well as the tool's individual commands) and
'git-bundle-web-server'.

In addition to the Asciidoc pages, add custom 'man:' and 'url:' macros in
'asciidoctor-extensions.rb'. The former is a simplified version of an
Asciidoctor example extension [1] (providing basic markup for references to
other man pages), while the latter overrides the default autolinking [2]
behavior in 'asciidoctor' to avoid some roff rendering weirdness and applies
some simple markup.

These files are intended for use with Asciidoctor. Documents can be
generated with:

$ asciidoctor -I scripts/ -r asciidoctor-extensions.rb docs/man/*.adoc

In a later commit, these docs will be built into installable manpages as
part of a 'Makefile' target.

[1] https://github.com/asciidoctor/asciidoctor-extensions-lab/blob/main/lib/man-inline-macro/extension.rb
[2] https://docs.asciidoctor.org/asciidoc/latest/macros/autolinks/

Signed-off-by: Victoria Dye <vdye@github.com>