filesystem: gracefully handle missing files in 'ReadFileLines()'
If the file argument to 'ReadFileLines()' does not exist on disk, return an
empty string list (corresponding to an empty file) with no error rather than
pass the error on to the caller.

This removes a pre-existing side-effect behavior of 'ReadFileLines()', where
the 'os.O_CREATE' flag would create the file if it was missing when read. It
would not create leading directories, though, leading to unhandled error
cases (e.g. in 'git-bundle-server init' without '~/git-bundle-server'). By
instead returning an empty result for any missing file, these errors are
avoided.

Signed-off-by: Victoria Dye <vdye@github.com>