log: add command logging
Add a 'LogCommand()' function to the 'TraceLogger', called in
'argparser.InvokeSubcommand()' right before the command is invoked. To
remain consistent with Git, do this only when a toplevel command (indicated
by 'isTopLevel') is invoked; nothing is done for subcommands of those
commands.

In the trace2 logger, this corresponds to the "cmd_name" event.

Signed-off-by: Victoria Dye <vdye@github.com>