cmd: add configurability to 'Run()'
Change the function signature of 'Run()' to accept command arguments in a
'[]string' (rather than variadic args) and add a set of 'cmd.Setting'
variadic args to optionally configure the command execution. This
configurability will be necessary for cutting over things like 'git.go'
(which, for example, sometimes specifies a custom stdin buffer) to use the
'CommandExecutor'.

To maintain the old function signatures, also add convenience wrappers
'RunQuiet()' and 'RunStdout()'.

Signed-off-by: Victoria Dye <vdye@github.com>