log: create initial trace2 logger
Create a general 'TraceLogger' interface and (for now) a sole implementation
in the 'Trace2' struct. There are two main components of this commit: the
Trace2 logger itself, and how it's used in the CLIs built in this
repository.

The Trace2 logger is built around the 'zap' structured logger. As of right
now, it does not log anything other than 'start', 'exit', and 'atexit' (and
only through the 'WithLogger()' wrapper described later). The logs include:

* 'event': the event type; it takes over the 'message' field of the 'zap'
  logging functions, mostly for the sake of simplicity (those functions
  require a message, and every trace2 log has an 'event').
* 'sid': a UUID generated by the 'uuid.NewUUID()' function, propagated
  through the call stack via the 'context.Context'.
* 'time': the UTC timestamp of the log generation; this is automatically
  provided by a 'zap' logger, but requires a custom formatter to provide the
  timestamp in UTC.
* 't_abs': the time (in seconds) since the start of the program
  (technically, when the 'log' package is loaded and the 'globalStartTime'
  is initialized).
* 'file' and 'line': the filename and line number where the relevant trace2
  log function was called.
* 'thread': hardcoded to 'main', but may be updated in the future to give
  unique names to goroutines.

From a user's perspective, the logger is configured by the
'GIT_TRACE2_EVENT' environment variable. Its behavior matches what is
described in Git [1], but it is currently missing support for open file
handles 2-9 and Unix Domain Sockets.

To use the trace2 logger, create the 'main()' function wrapper
'WithLogger()'. This function creates an instance of a 'TraceLogger', logs
the program start, defers a "cleanup" logging function to write out the
'exit'/'atexit' logs, then runs the wrapped function.

[1] https://git-scm.com/docs/git-config#Documentation/git-config.txt-trace2eventTarget

Signed-off-by: Victoria Dye <vdye@github.com>