log: add error logging
Add logging for errors returned 'fmt.Errorf()' errors as well as the errors
passed to or created by 'TraceLogger.Fatal()'/'TraceLogger.Fatalf()'.
Replace all use of 'fmt.Errorf()' or direct error returns with
'TraceLogger.Errorf()' or 'TraceLogger.Error()', which return 'error'
instances to stay consistent with what they're replacing.

These errors are logged under the "error" trace2 event. We want to capture
the error as low in the call stack as possible (to more precisely indicate
the cause), but also not log a new "error" event for every 'fmt.Errorf()'
replaced. To do that, have the error wrappers return a custom type
'loggedError', and only log the error event if the 'err' or none of the
'Errorf()' arguments are of the 'loggedError' type.

Signed-off-by: Victoria Dye <vdye@github.com>