testhelpers: create MockTraceLogger
Create a 'MockTraceLogger' struct that, like other mocks in 'mock.go',
embeds a 'mock.Mock' struct. However, unique to the 'TraceLogger' instance
in tests (namely the daemon tests) is that we usually don't want to assert
on the logger's behavior; in most cases, we just want it to "do nothing" or
pass through.

Implement the 'TraceLogger' interface with checks that only invoke
'Mock.Called()' if the method has been mocked at least once (even if it's
for args not matching the current ones). If the method is not mocked, fall
back on "passthrough" defaults (usually returning the input
'context.Context' and/or the input args).

Finally, add an input assertion to 'MockTraceLogger.Error()', to serve as an
additional validation in unit tests and mirror the behavior of an actual
'TraceLogger' instance.

Signed-off-by: Victoria Dye <vdye@github.com>