filesystem: add 'GetLocalExecutable()'
Add a function 'GetLocalExecutable()' to 'FileSystem'. This function, given
the identifier of one of the executables installed with this package
('git-bundle-server' or 'git-bundle-web-server'), will look for the file in
the same directory as the current executing program, and return its absolute
path.

This function is then used to replace the custom logic in
'getDaemonConfig()', which locates the 'git-bundle-web-server' executable to
run in a daemon process. This is a functional change; previously,
'getDaemonConfig()' would first lookup the executable with 'exec.LookPath()'
before falling back on the local directory. Removing the 'exec.LookPath()'
seems like the right thing to do here - if we're running a local version of
'git-bundle-server', we'd want to find the executable corresponding to that
local build.

Signed-off-by: Victoria Dye <vdye@github.com>