utils: add reflection-driven dependency container
Add a 'DependencyContainer' type, which contains methods to register and
lazily retieve singleton dependencies.

This is implemented in anticipation of a substantial increase in the
complexity of the dependency hierarchy. Its primary benefit is avoiding the
need to initialize an entire hierarchy of dependent structs "manually",
hopefully making for easier to read (and modify) code. Note, though, that
this design carries some risk:

- errors due to unregistered components will only manifest at runtime.
- we might see potentially slower startup due to the use of reflection; that
  said, in manual testing, I didn't notice a slowdown.
- we can end up in a recursive loop of dependency resolution, ultimately
  leading to a stack overflow.

To mitigate that risk, add a test 'TestBuildGitBundleServerContainer', which
uses a combination of reflection and (somewhat fragile, but documented in
the test) abstract syntax tree parsing to verify that the container is
fully-registered.

Signed-off-by: Victoria Dye <vdye@github.com>