bundles: add region tracing
Add 'TraceLogger.Region()' invocations to all public methods of the
'BundleProvider' struct, except for 'CreateInitialBundle()' and
'CreateSingletonBundle()' (due to the relative brevity of those functions).

Note that static check suppressions are added to two of the 'Region()' calls
to allow overriding the 'ctx' despite it not being used later in the
function. This helps future-proof against the easy-to-miss bug where a call
is added to the function that uses 'ctx', but because 'ctx' is not
overridden with the value from 'Region()', the ctx doesn't include the
appropriate region information.

Signed-off-by: Victoria Dye <vdye@github.com>