release.yml: add signing for MacOS artifacts
Add a step to set up the signing keychain for signing the MacOS package
contents with the Apple Developer certificates, and enable
signing/notarization in the 'make package' build that generates the
artifacts.

In order to make the signing process more secure, the MacOS packaging jobs
are performed in a 'release' environment.

Note that it is possible to run the workflow with no signing/notarization,
signing-only (no notarization), or both signing & notarization; the behavior
depends on which workflow secrets are set. This ensures that artifacts can
be successfully built in a variety of environments, making it easier for
forks to test the release build should they want to.

Signed-off-by: Victoria Dye <vdye@github.com>