git: capture Git STDERR log in Go error
In a609c0b (git: refactor into interface, 2023-02-24), the function
'GitCommandWithStdin()' was replaced with 'gitCommandWithStdin()'. In that
update, the STDERR log that was previously written to the error message was
removed. Downstream functions (namely 'CreateIncrementalBundle()') relied on
the content of that STDERR message to determine behavior. Add it back to the
error message to correct the behavor of 'CreateIncrementalBundle()'.

Additionally, add unit tests covering 'CreateIncrementalBundle()'.

Signed-off-by: Victoria Dye <vdye@github.com>