filesystem: add lockfile write function
Add 'WriteLockFileFunc()' to provide a mechanism for updating files
semi-atomically. The function takes a filename and a function (the logic
perform the write to a given file pointer, executes the write function on
'<filename>.lock', then returns a 'LockFile' object. That returned
'LockFile' can then either invoke 'Commit()' (to rename '<filename>.lock' to
'<filename>') or 'Rollback()' (to delete '<filename>.lock'). If
'WriteLockFileFunc()' encounters any error in its operation, it tries to
rollback the lockfile before returning the error.

This function is mainly useful for writing files that may be read
concurrently with the write operation. Writing the file to a lockfile, then
atomically renaming the file avoids avoids the possibility of another
process reading a half-written update to the file.

Example usage:

---
lockFile, err := fileSystem.WriteLockFileFunc("myfile.txt", func(f io.Writer) error {
    f.Write([]byte("my text"))
})
if err != nil {
    return err
}
lockFile.Commit()
---

Signed-off-by: Victoria Dye <vdye@github.com>