bundle-server: serve with 'ServeContent()'
Use 'OpenFile()' + 'ServeContent()', rather than 'ReadFile()' + 'Write()' to
serve bundle and bundle list content. There are a few (fairly minor)
benefits to 'ServeContent()': it does not buffer the full file contents
before writing to the response writer, and it automatically derives the
'Content-Type' and other headers from the file content.

Signed-off-by: Victoria Dye <vdye@github.com>