bundles: write bundle lists for different URIs
The bundle list for a given 'git-bundle-web-server' is intended to be served
from the route 'http[s]://<host>:<port>/<org>/<repo>', with or without a
trailing slash. The bundles listed are provided as paths relative to
'/<org>/<repo>'. However, these relative paths only work when there's a
trailing slash after '<repo>' in the URI; otherwise '<repo>' is interpreted
as a file in directory '<org>'. For example:

- URI: https://localhost:8080/git/git/, bundle relpath: bundle-1.bundle
  -> https://localhost:8080/git/git/bundle-1.bundle
- URI: https://localhost:8080/git/git, bundle relpath: bundle-2.bundle
  -> https://localhost:8080/git/bundle-2.bundle

This behavior is baked into Git, and is consistent with RFC 1808 [1]. We
still want to be able to serve the bundle list when no trailing slash is
given, though, so write out a second bundle list that prepends '<repo>' to
the bundle filename. In a later commit, we will update the web server to
serve this file's contents when appropriate, fixing the routing bug.

As part of this refactor of 'BundleProvider.WriteBundleList()', switch to
using 'FileSystem.WriteLockFileFunc()' to write lock files. Doing so allows
us to mock filesystem writes and unit test 'WriteBundleList()', which is
done here as well.

[1] https://www.rfc-editor.org/rfc/rfc1808#section-4

Signed-off-by: Victoria Dye <vdye@github.com>