bundle-server: serve correct bundle routes for request
Update 'bundleServer.serve()' to serve the correct bundle list file for the
request URI. If the URI has no trailing slash (e.g.
'https://<host>:<port>/<org>/<repo>'), serve the 'repo-bundle-list' file
contents, which prepends '<repo>' to the relative bundle paths. If the URI
has a trailing slash, serve 'bundle-list' as normal.

Also, prevent users from directly reading 'bundle-list' and
'repo-bundle-list' based on the URI; the latter's relative paths would be
incorrect, and in general these files are an internal implementation detail
that should not be visible to users. If a user requests one of these files,
return a 404.

Signed-off-by: Victoria Dye <vdye@github.com>