workflows: use go version specified in 'go.mod'
Update the workflows that build the project to use the fixed version of Go
tied to the project in 'go.mod'. The 'go.mod' can be used as the sole source
of truth on the supported version of Go in the repository, making it easier
to upgrade and use new features when needed.

Signed-off-by: Victoria Dye <vdye@github.com>