init: parse route from URL
Make the '<route>' positional arg optional in 'git-bundle-server init'. If
it is not specified (or is an empty string), extract a route from the
supplied clone URI.

Called by 'init', 'core.GetRouteFromUrl()' attempts to match the URL against
SSH, HTTP(S), and filesystem ('file://') URLs. If the URL cannot be matched,
'init' will exit with an error and usage printout.

Also include unit tests of the route matching function to demonstrate its
behavior.

Signed-off-by: Victoria Dye <vdye@github.com>