git: add 'GetRemoteUrl()'
Add a 'GetRemoteUrl()' function to 'git.GitHelper', which calls 'git remote
get-url origin' on a given repository path and returns the whitespace-
trimmed remote URL. Neither stdout nor stderr are printed to the console,
but the stderr text is included in the error message if Git exits with an
error status.

Although it is not used in this patch, 'GetRemoteUrl()' will be used in
later commits to display information about the configuration of the bundle
server.

Signed-off-by: Victoria Dye <vdye@github.com>