repo: add 'ReadRepositoryStorage()'
Add a function to the 'core.RepositoryProvider' that reads the on-disk
repository storage for the bundle server and returns the valid repos in a
route -> 'core.Repository' map.

'ReadRepositoryStorage()' identifies repositories by finding the directories
at a depth of 2 from the repo storage root, then checking that the
repository has a remote 'origin'. The result includes the "active" routes in
the bundle server (stored in the 'routes' file) as well as those that have
been removed from the routes file with 'git-bundle-server stop'. In later
commits, this information will be used to provide more complete information
about the state of the bundle server and configure "repairs" to a corrupted
server.

Signed-off-by: Victoria Dye <vdye@github.com>