filesystem: remove unused mock function
Remove mock of 'UserHomeDir()' for the 'FileSystem' interface leftover from
an earlier iteration of de58f24 (repo: add more GetRepositories() tests, fix
bug, 2023-01-31). The function was removed by the time the commit was
merged, so the mock can safely be deleted.

Add a call to 'mock.AssertExpectationsForObjects()' to ensure all mocked
functions are called. This function should be called in all mocked unit
tests to ensure similar issues do not occur in the future.

Signed-off-by: Victoria Dye <vdye@github.com>