update: improve informational printouts
Add and update informational printouts in 'git-bundle-server update' to
inform users more accurately of which step is in progress, as well as
clearly indicate when the update stops early due to a lack of updates from
the remote.

Also add a printout to 'update-all' indicating which route is being updated
at a given time as it loops through all routes.

Signed-off-by: Victoria Dye <vdye@github.com>