git-bundle-server: add 'list' command
Add a new 'git-bundle-server' command to list the routes that are currently
configured and their remote Git URLs. Update manpage and 'README.md' with
description and usage information.

Due to how routes are currently stored in the bundle server, only "active"
routes are listed (that is, routes that have not been stopped with
'git-bundle-server stop'). This is intended to be only a temporary issue,
with future updates to how routes are stored allowing better tracking of
"stopped" routes.

Signed-off-by: Victoria Dye <vdye@github.com>