bundle-server: add client cert verification
Add a '--client-ca' option to 'git-bundle-web-server' that, when specified,
loads a PEM-formatted certificate file and uses the cert(s) within to verify
client requests to the bundle web server. If the bundle server has been
configured with this option and the client does not send an appropriate
certificate, the TLS handshake will fail and the client will be unable to
connect.

Configuring both server TLS hosting (via the '--cert' and '--key' options)
and client cert validation (via '--client-ca') sets up mutual TLS (mTLS)
authentication on the bundle web server.

Signed-off-by: Victoria Dye <vdye@github.com>