e2e tests: updates for integration tests
Update necessary components of existing e2e tests to prepare for new
integration test suite. This includes:

1. Moving package-related files (package-lock.json, package.json, and
tsconfig.json) out of the 'e2e' directory and into 'test' and updating
package.json to not be e2e-test specific. This will allow both suites to
be combined into a single Typescript package in upcoming patches.i

2. Moving certain classes currently in the e2e 'step_definitions',
'support', and 'classes' directories that will be needed by integration
tests into a new 'shared' directory. This led to notable refactorings to
implement the shared 'BundleServerWorld' class as a parent and a new
'E2EBundleServerWorld' class as a child. A similar
'IntegrationBundleServerWorld' class will be added in a future patch.

3. Updating the 'require' statement in the e2e tests' config to identify
the correct 'features' directories with the new architecture. This
indicates to Cucumber that it should look in multiple directories for
features. The VSCode settings have been updated to recognize these
directories as well.

Signed-off-by: Lessley Dennington <lessleydennington@gmail.com>